Curl parts download test
Was playing around with curl trying to see if it was faster to break the file up into five parts and download it with different curl connection, then reassemble the file. In this test it equaled roughly the same time as a simple curl command. Maybe take another look later.